Moved some other function to the new util.
