+++ src/main/java/xyz/alexcrea/cuanvil/gui/util/GuiSharedConstant.java
@@ -9,7 +9,8 @@ import org.bukkit.inventory.ItemStack;
 import org.bukkit.inventory.meta.ItemMeta;
 import xyz.alexcrea.cuanvil.gui.config.MainConfigGui;
 
-import java.util.*;
+import java.util.Arrays;
+import java.util.Collections;
 
 public class GuiSharedConstant {
